+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2025-11-28
+
 ### Added
 
 - Support `allOf`, `oneOf`, `const` and `enum`
@@ -70,7 +72,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/giantswarm/schemadocs/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/giantswarm/schemadocs/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/giantswarm/schemadocs/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/giantswarm/schemadocs/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/giantswarm/schemadocs/compare/v0.0.7...v0.1.0
 [0.0.7]: https://github.com/giantswarm/schemadocs/compare/v0.0.6...v0.0.7